+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/Chat/OnlineAsyncTaskAccelByteChatGetSystemMessagesStats.cpp
@@ -53,7 +53,7 @@ void FOnlineAsyncTaskAccelByteChatGetSystemMessagesStats::TriggerDelegates()
 void FOnlineAsyncTaskAccelByteChatGetSystemMessagesStats::OnGetSystemMessagesStatsSuccess(
 	const FAccelByteGetSystemMessageStatsResponse& Response)
 {
-	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("UserId: %d"), *UserId->GetAccelByteId());
+	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("UserId: %s"), *UserId->GetAccelByteId());
 
 	SystemMessageStats = Response;
 	OnlineError = ONLINE_ERROR(EOnlineErrorResult::Success, FString(), FText());
@@ -65,7 +65,7 @@ void FOnlineAsyncTaskAccelByteChatGetSystemMessagesStats::OnGetSystemMessagesSta
 void FOnlineAsyncTaskAccelByteChatGetSystemMessagesStats::OnGetSystemMessagesStatsError(int32 ErrorCode,
 	const FString& ErrorMessage)
 {
-	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("UserId: %d"), *UserId->GetAccelByteId());
+	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("UserId: %s"), *UserId->GetAccelByteId());
 	OnlineError = ONLINE_ERROR(EOnlineErrorResult::RequestFailure, FString::FromInt(ErrorCode), FText::FromString(ErrorMessage));
 	CompleteTask(EAccelByteAsyncTaskCompleteState::RequestFailed);
 	AB_OSS_ASYNC_TASK_TRACE_END(TEXT("Failed to query system messages, ErrorCode: %d, ErrorMessage: %s"), ErrorCode, *ErrorMessage);